Move legacy files to legacy folder
