+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_condition_evaluations.py
@@ -4,7 +4,7 @@ from typing import Optional, Sequence, Union
 
 import graphene
 from dagster._core.definitions.asset_subset import AssetSubset
-from dagster._core.definitions.declarative_scheduling.asset_condition import (
+from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import (
     AssetConditionEvaluation,
 )
 from dagster._core.definitions.partition import PartitionsDefinition, PartitionsSubset
