+++ python_modules/dagster/dagster/_core/definitions/asset_daemon_cursor.py
@@ -28,7 +28,7 @@ from dagster._serdes.serdes import (
 from .base_asset_graph import BaseAssetGraph
 
 if TYPE_CHECKING:
-    from .declarative_scheduling.asset_condition import (
+    from .declarative_scheduling.legacy.asset_condition import (
         AssetConditionEvaluation,
         AssetConditionEvaluationState,
         AssetConditionSnapshot,
@@ -149,10 +149,10 @@ def get_backcompat_asset_condition_evaluation_state(
 ) -> "AssetConditionEvaluationState":
     """Generates an AssetDaemonCursor from information available on the old cursor format."""
     from dagster._core.definitions.auto_materialize_rule_impls import MaterializeOnMissingRule
-    from dagster._core.definitions.declarative_scheduling.asset_condition import (
+    from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import (
         AssetConditionEvaluationState,
     )
-    from dagster._core.definitions.declarative_scheduling.operators.rule_operator import (
+    from dagster._core.definitions.declarative_scheduling.legacy.rule_condition import (
         RuleCondition,
     )
 
@@ -178,7 +178,7 @@ def backcompat_deserialize_asset_daemon_cursor_str(
     from .auto_materialize_rule_evaluation import (
         deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_with_run_ids,
     )
-    from .declarative_scheduling.asset_condition import (
+    from .declarative_scheduling.legacy.asset_condition import (
         AssetConditionEvaluation,
         AssetConditionSnapshot,
     )
