+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_evaluation.py
@@ -34,7 +34,7 @@ from dagster._utils.security import non_secure_md5_hash_str
 from .partition import PartitionsDefinition, SerializedPartitionsSubset
 
 if TYPE_CHECKING:
-    from .declarative_scheduling.asset_condition import (
+    from .declarative_scheduling.legacy.asset_condition import (
         AssetConditionEvaluation,
         AssetConditionEvaluationWithRunIds,
         AssetConditionSnapshot,
@@ -148,7 +148,9 @@ def deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_
     """Provides a backcompat layer to allow deserializing old AutoMaterializeAssetEvaluation
     objects into the new AssetConditionEvaluationWithRunIds objects.
     """
-    from .declarative_scheduling.asset_condition import AssetConditionEvaluationWithRunIds
+    from .declarative_scheduling.legacy.asset_condition import (
+        AssetConditionEvaluationWithRunIds,
+    )
 
     class BackcompatDeserializer(BackcompatAutoMaterializeAssetEvaluationSerializer):
         @property
@@ -213,8 +215,8 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
     def _asset_condition_snapshot_from_rule_snapshot(
         self, rule_snapshot: AutoMaterializeRuleSnapshot
     ) -> "AssetConditionSnapshot":
-        from .declarative_scheduling.asset_condition import AssetConditionSnapshot
-        from .declarative_scheduling.operators.rule_operator import RuleCondition
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionSnapshot
+        from .declarative_scheduling.legacy.rule_condition import RuleCondition
 
         unique_id_parts = [rule_snapshot.class_name, rule_snapshot.description]
         unique_id = non_secure_md5_hash_str("".join(unique_id_parts).encode())
@@ -234,7 +236,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         rule_snapshot: AutoMaterializeRuleSnapshot,
     ) -> "AssetConditionEvaluation":
-        from .declarative_scheduling.asset_condition import (
+        from .declarative_scheduling.legacy.asset_condition import (
             AssetConditionEvaluation,
             AssetSubsetWithMetadata,
             HistoricalAllPartitionsSubsetSentinel,
@@ -279,7 +281,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         decision_type: AutoMaterializeDecisionType,
     ) -> Optional["AssetConditionEvaluation"]:
-        from .declarative_scheduling.asset_condition import (
+        from .declarative_scheduling.legacy.asset_condition import (
             AssetConditionEvaluation,
             AssetConditionSnapshot,
             HistoricalAllPartitionsSubsetSentinel,
@@ -374,7 +376,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         whitelist_map: WhitelistMap,
         context: UnpackContext,
     ) -> "AssetConditionEvaluationWithRunIds":
-        from .declarative_scheduling.asset_condition import (
+        from .declarative_scheduling.legacy.asset_condition import (
             AssetConditionEvaluation,
             AssetConditionSnapshot,
             HistoricalAllPartitionsSubsetSentinel,
