+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_impls.py
@@ -47,7 +47,7 @@ from dagster._utils.schedules import (
 from .base_asset_graph import sort_key_for_asset_partition
 
 if TYPE_CHECKING:
-    from dagster._core.definitions.declarative_scheduling.asset_condition import (
+    from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import (
         AssetConditionResult,
     )
 
@@ -75,7 +75,7 @@ class MaterializeOnRequiredForFreshnessRule(
         return "required to meet this or downstream asset's freshness policy"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         true_subset, subsets_with_metadata = freshness_evaluation_results_for_asset_key(
             context.legacy_context.root_context
@@ -189,7 +189,7 @@ class MaterializeOnCronRule(
             }
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         missed_ticks = self.missed_cron_ticks(context)
         new_asset_partitions = self.get_new_candidate_asset_partitions(context, missed_ticks)
@@ -345,7 +345,7 @@ class MaterializeOnParentUpdatedRule(
         """Evaluates the set of asset partitions of this asset whose parents have been updated,
         or will update on this tick.
         """
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         asset_partitions_by_updated_parents: Mapping[
             AssetKeyPartitionKey, Set[AssetKeyPartitionKey]
@@ -470,7 +470,7 @@ class MaterializeOnMissingRule(AutoMaterializeRule, NamedTuple("_MaterializeOnMi
         """Evaluates the set of asset partitions for this asset which are missing and were not
         previously discarded.
         """
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         if (
             context.legacy_context.asset_key
@@ -549,7 +549,7 @@ class SkipOnParentOutdatedRule(AutoMaterializeRule, NamedTuple("_SkipOnParentOut
         return "waiting on upstream data to be up to date"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -602,7 +602,7 @@ class SkipOnParentMissingRule(AutoMaterializeRule, NamedTuple("_SkipOnParentMiss
         self,
         context: "SchedulingContext",
     ) -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -678,7 +678,7 @@ class SkipOnNotAllParentsUpdatedRule(
         self,
         context: "SchedulingContext",
     ) -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -903,7 +903,7 @@ class SkipOnNotAllParentsUpdatedSinceCronRule(
             )
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         passed_time_window = self.passed_time_window(context)
         has_new_passed_time_window = passed_time_window.end.timestamp() > (
@@ -977,7 +977,7 @@ class SkipOnRequiredButNonexistentParentsRule(
         return "required parent partitions do not exist"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -1028,7 +1028,7 @@ class SkipOnBackfillInProgressRule(
             return "targeted by an in-progress backfill"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         backfilling_subset = (
             # this backfilling subset is aware of the current partitions definitions, and so will
@@ -1063,7 +1063,7 @@ class DiscardOnMaxMaterializationsExceededRule(
         return f"exceeds {self.limit} materialization(s) per minute"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         # the set of asset partitions which exceed the limit
         rate_limited_asset_partitions = set(
@@ -1094,7 +1094,7 @@ class SkipOnRunInProgressRule(AutoMaterializeRule, NamedTuple("_SkipOnRunInProgr
         return "in-progress run for asset"
 
     def evaluate_for_asset(self, context: "SchedulingContext") -> "AssetConditionResult":
-        from .declarative_scheduling.asset_condition import AssetConditionResult
+        from .declarative_scheduling.legacy.asset_condition import AssetConditionResult
 
         if context.legacy_context.partitions_def is not None:
             raise DagsterInvariantViolationError(
