+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/legacy/rule_condition.py
@@ -5,8 +5,8 @@ from dagster._core.definitions.auto_materialize_rule import AutoMaterializeRule
 from dagster._serdes.serdes import whitelist_for_serdes
 from dagster._utils.security import non_secure_md5_hash_str
 
-from ..asset_condition import AssetCondition, AssetConditionResult
 from ..scheduling_context import SchedulingContext
+from .asset_condition import AssetCondition, AssetConditionResult
 
 
 @experimental
