+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/operands/updated_since_cron_condition.py
@@ -2,7 +2,7 @@ import datetime
 
 from dagster._utils.schedules import reverse_cron_string_iterator
 
-from ..asset_condition import AssetCondition, AssetConditionResult
+from ..legacy.asset_condition import AssetCondition, AssetConditionResult
 from ..scheduling_context import SchedulingContext
