+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/custom_condition_scenarios.py
@@ -1,7 +1,7 @@
 from dagster import AutoMaterializeRule
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
-from dagster._core.definitions.declarative_scheduling.asset_condition import AssetCondition
-from dagster._core.definitions.declarative_scheduling.operators.rule_operator import RuleCondition
+from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import AssetCondition
+from dagster._core.definitions.declarative_scheduling.legacy.rule_condition import RuleCondition
 
 from ..base_scenario import run_request
 from ..scenario_specs import one_asset, two_assets_in_sequence
